+++ sources/@roots/bud/src/cli/commands/bud.view.tsx
@@ -59,7 +59,7 @@ export default class BudViewCommand extends BudCommand {
 
     if (this.color) value = highlight(value)
 
-    this.renderOnce(
+    this.renderStatic(
       <Ink.Box>
         <Ink.Text color="magenta">
           {this.subject ?? `build.config`}
