+++ sources/@roots/bud/src/cli/commands/doctor/Doctor.tsx
@@ -1,0 +1,51 @@
+/* eslint-disable react/no-unescaped-entities */
+import figures from '@roots/bud-support/figures'
+import * as Ink from 'ink'
+
+import {Error} from '../../components/Error.js'
+import {isWindows} from '../../helpers/isWindows.js'
+
+export const Doctor = ({name, timings}) => {
+  return (
+    <Ink.Box flexDirection="column" marginTop={1}>
+      <Ink.Text underline>{`Diagnosis for ${name}\n`}</Ink.Text>
+      <Ink.Text dimColor>
+        Completed a dry run of your project's build (executed in{` `}
+        {timings.build} seconds). If the information provided by this
+        command doesn't yield a solution consider running `yarn bud repl`
+        and exploring the finalized config (`bud.build.config`).
+      </Ink.Text>
+      <Process />
+    </Ink.Box>
+  )
+}
+
+const Process = () => {
+  return (
+    <Ink.Box marginTop={1} flexDirection="column">
+      <Ink.Text color="blue">Checking system requirements{`\n`}</Ink.Text>
+
+      <Ink.Box flexDirection="column">
+        <Ink.Text>
+          {process.version.match(/v1[6|7|8|9]/)
+            ? figures.tick
+            : figures.cross}
+          {` `}
+          node: {process.version}
+        </Ink.Text>
+
+        <Ink.Text>
+          {isWindows() ? figures.cross : figures.tick} os:{` `}
+          {process.platform}
+        </Ink.Text>
+
+        {!process.version.match(/v1[6|7|8|9]/) && (
+          <Error
+            name="Node version not supported"
+            message={`Please upgrade to Node v18 for long-term support. You are running node ${process.version}.`}
+          />
+        )}
+      </Ink.Box>
+    </Ink.Box>
+  )
+}
